# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "نظام إدارة المكتبة الشخصية وتقييم الكتب (Personal Library API)",
        "description": "RESTful API متكامل لإدارة الكتب، المؤلفين، التصنيفات، وتقييمات الكتب مع توثيق كامل باستخدام Swagger.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "سيرفر التطوير المحلي (Local Development Server)"
        }
    ],
    "paths": {
...
        "/authors": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "استعراض كافة المؤلفين",
                "operationId": "32b5789d7dff4977d5cca5fe881af025",
                "responses": {
                    "200": {
                        "description": "قائمة المؤلفين",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "نجيب محفوظ"
                                            },
                                            "bio": {
                                                "type": "string",
                                                "example": "روائي مصري حائز على جائزة نوبل في الأدب."
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Authors"
                ],
                "summary": "إضافة مؤلف جديد",
                "operationId": "561c1072fa32b9ddedf8d8ecff87b0c7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "نجيب محفوظ"
                                    },
                                    "bio": {
                                        "type": "string",
                                        "example": "روائي مصري حائز على جائزة نوبل في الأدب."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "تمت إضافة المؤلف بنجاح",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "نجيب محفوظ"
                                        },
                                        "bio": {
                                            "type": "string",
                                            "example": "روائي مصري حائز على جائزة نوبل في الأدب."
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "غير مصرح (Unauthorized)"
                    },
                    "422": {
                        "description": "الاسم مطلوب"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/authors/{id}": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "عرض تفاصيل مؤلف محدد",
                "operationId": "739e4ae62699b259f765b5c8385277b1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "معرّف المؤلف",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "تفاصيل المؤلف",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "نجيب محفوظ"
                                        },
                                        "bio": {
                                            "type": "string",
                                            "example": "روائي مصري حائز على جائزة نوبل في الأدب."
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "المؤلف غير موجود"
                    }
                }
            },
            "put": {
                "tags": [
                    "Authors"
                ],
                "summary": "تحديث معلومات مؤلف",
                "operationId": "3ef55162226f76a98a1687aacb3d4529",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "معرّف المؤلف المراد تحديثه",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "طه حسين"
                                    },
                                    "bio": {
                                        "type": "string",
                                        "example": "عميد الأدب العربي."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "تم تحديث بيانات المؤلف بنجاح",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "طه حسين"
                                        },
                                        "bio": {
                                            "type": "string",
                                            "example": "عميد الأدب العربي."
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "غير مصرح"
                    },
                    "404": {
                        "description": "المؤلف غير موجود"
                    },
                    "422": {
                        "description": "بيانات التحقق غير صالحة"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Authors"
                ],
                "summary": "حذف مؤلف",
                "operationId": "203198cefe94cb4fd5def3d8e42bb87c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "معرّف المؤلف المراد حذفه",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "تم حذف المؤلف بنجاح",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "تم حذف المؤلف بنجاح"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "غير مصرح"
                    },
                    "404": {
                        "description": "المؤلف غير موجود"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "أدخل الـ Token الخاص بك هنا بالصيغة التالية: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}